refactor: 컨트롤러 병합
