+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -38,4 +38,14 @@ public class LoginController {
             return "redirect:/401.jsp";
         }
     }
+
+    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
+    public String loginView(final HttpServletRequest req, final HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return "redirect:/index.jsp";
+                })
+                .orElse("/login.jsp");
+    }
 }
